[IMP] bus: convert im_status tests to hoot
preparation for PR: https://github.com/odoo/odoo/pull/177611
