+++ addons/mail/static/tests/mail_test_helpers_contains.js
@@ -5,6 +5,9 @@ import { queryAll, queryFirst } from "@odoo/hoot-dom";
 import { Deferred, tick } from "@odoo/hoot-mock";
 import { isMacOS } from "@web/core/browser/feature_detection";
 import { isVisible } from "@web/core/utils/ui";
+import { step as webStep, assertSteps as webAssertSteps } from "@web/../tests/web_test_helpers";
+export const step = webStep;
+export const assertSteps = webAssertSteps;
 
 /** @param {EventInit} [args] */
 const mapBubblingEvent = (args) => ({ ...args, bubbles: true });
@@ -928,74 +931,4 @@ class Contains {
  */
 export async function contains(selector, options) {
     await new Contains(selector, options).run();
-}
-
-const stepState = {
-    expectedSteps: null,
-    /** @type {Promise} */
-    deferred: null,
-    timeout: null,
-    currentSteps: [],
-
-    clear() {
-        clearTimeout(this.timeout);
-        this.timeout = null;
-        this.deferred = null;
-        this.currentSteps = [];
-        this.expectedSteps = null;
-    },
-
-    check({ crashOnFail = false } = {}) {
-        const success =
-            this.expectedSteps.length === this.currentSteps.length &&
-            this.expectedSteps.every((s, i) => s === this.currentSteps[i]);
-        if (!success && !crashOnFail) {
-            return;
-        }
-        expect.verifySteps(this.expectedSteps);
-        if (success) {
-            this.deferred.resolve();
-        } else {
-            this.deferred.reject(new Error("Steps do not match."));
-        }
-        this.clear();
-    },
-};
-
-afterEach(() => {
-    if (stepState.expectedSteps) {
-        stepState.check({ crashOnFail: true });
-    } else {
-        stepState.clear();
-    }
-});
-
-/**
- * Indicate the completion of a test step. This step must then be verified by
- * calling `assertSteps`.
- *
- * @param {string} step
- */
-export function step(step) {
-    stepState.currentSteps.push(step);
-    expect.step(step);
-    if (stepState.expectedSteps) {
-        stepState.check();
-    }
-}
-
-/**
- * Wait for the given steps to be executed or for the timeout to be reached.
- *
- * @param {string[]} steps
- */
-export function assertSteps(steps) {
-    if (stepState.expectedSteps) {
-        stepState.check({ crashOnFail: true });
-    }
-    stepState.expectedSteps = steps;
-    stepState.deferred = new Deferred();
-    stepState.timeout = setTimeout(() => stepState.check({ crashOnFail: true }), 2000);
-    stepState.check();
-    return stepState.deferred;
 }
